CSS import: added new function 'adjustLengthValues ($callback)'. The function can be used to do a specific length conversion of every imported CSS property. The callback function $callback will be called for every property which carries a length value. The value will then be set to the value that the callback function returns.
The callback is given the following parameters:
$property: The name of the property, e.g. 'width'
$value: It's value from the imported CSS, e.g. '300px'
$type: One of...
  CSSValueType::LengthValueXAxis : The value is a length on the X axis
  CSSValueType::LengthValueYAxis : The value is a length on the Y axis
  CSSValueType::StrokeOrBorderWidth: The value is a stroke or border width etc..

The ODT renderer uses the function to not convert all pixel based length values the same
way, but in a differentiated way. By examing the $type, the callback function can convert
X values in a different way then Y values. Or it can do more specific things based on the
$property itself.