Only delete empty paragraphs if they are really useless. Paragraphs used for a pagebreak or changing page format may not be deleted. Fixes #58.
To decide if an empty paragraph may be deleted or not the style-name is being checked.
If the style name references a pagebreak style or a page format style, then the paragraph is not deleted.
All style-names that prevent deletion of empty paragraphs are listed in the array $preventDeletetionStyles [].
If further paragraph styles - which may not be deleted - are added in the future, then they just need to be
added to the array to prevent unwanted deletion.