Implemented style classes ODTMasterPageStyle and ODTPageLayoutStyle. Before, the style elements 'style:master-page' and 'style:page-layout' had been handled as unknown styles (as objects of class ODTUnknownStyle).
This fixes setting of the start page format which did not work anymore since
re-designing of the style handling.

Also the page format of an ODT template file is now overwritten with the page format
set by the ODT plugin. The rest of the parameters from the templates page-layout are kept.
Only the page format is overwritten (format, orientation and margins).

Fixes #104, #108.