Added new functions to let plugins create their own styles for later use:
- createTextStyle($properties, $common)
- createParagraphStyle($properties, $common),
- createTableStyle($properties, $common)
- createTableRowStyle($properties, $common)
- createTableCellStyle($properties, $common)
- createTableColumnStyle($properties, $common)
- styleExists($style_name)

If $common is set to 'true' (default value) then the style will be added
to the common styles otherwise it will be added to the automatic styles.
Automatic styles are generated on direct formating of some content.
Common styles are visible and editable for the user e.g. in LibreOffice under 'Styles and Formatting'.

Therefore a common style should have these minimim properties:
- $properties ['style-name'] = 'Plugin_A_StyleName_A';
  The 'style-name' must be unique and is used in later function calls to identify the style.
  So a plugin should add it's name as a prefix to the style name to prevent clashes.
- $properties ['style-display-name'] = 'Some readable, meaningful name';
  The 'style-display-name' is the name that is presented to the user.
  It does not need to be unique but should have a meaningful and readable name.