Improved CSS support. This required new classes to not break the compatibility with existing functions/classes.
The new classes are placed in cssimportnew.php.
The functions in cssdocument.php are required to use the new CSS import
but they are independet from it. The classes is cssimportnew.php
only need to be given an object implementing the interface in
ecm_interface.php to work.

The main reason for this is that the old CSS import was not able
to properly match all kind of selectors.