Moved list code from ODTDocument.php to ODTList.php.
Why that?
ODTDocument.php is already becoming large and not nearly all ODT specific code from the renderer
has been moved yet. So I decided to move every element/functionality specific code to it's own static class.
In that way ODTDocument as the supposed API entry point will be represented clearly and will simply delegate
functionality to the corresponding static classes.

The following rules for the static classes like ODTList.php shall aply:
- all functions shall be static
- they shall not store any data themselves. An instance of ODTDocument shall be passed as argument 1 to each
  function to store data in it.

This also moves functionality/element specific, internal code away from the API in ODTDocument.php.
E.g. the function 'replaceLastListParagraph()' is only called by list code and has now been moved away from
ODTDocument and is not included there. It was not called from the outside anyway so I think it is good to not
have it in the "first level" code of the API.