Moved functionality of class 'helper_plugin_odt_cssimportnew' to class 'cssimportnew'.
The new class 'cssimportnew' is now located in 'ODT/css/cssimportnew.php' and is independent from DokuWiki.
That means the class does not extend/inherit from DokuWiki_Plugin.
The old class has been re-written to be a simple wrapper class for the new class.

So in DokuWiki context the helper plugin can still be used.
In the internal ODT context the DokuWiki independent class will be used.