Moved code for indexes (e.g. table of contents) to ODTDocument.php ==> ODTIndex.php.
The code has also been revised:
The settings for generating an index are passed as a multi dimensional array not as previously as multiple arrays.
In renderer/page.php in function 'get_index_settings()' the settings array for an index is generated out of the
syntax tag data and configuration data.