Copy properties of 'body' style to default paragraph style.
This solves the issue with dynamically created styles not having the correct font-size in case no
font-size was specified on creation, e.g. font-size in the Table of Contents.

Margins and paddings are not copied to the default style as this would break the layout
of the Table of Contents and lists.