Revised conversion of Geshi code to ODT format. Fixes #195.
Source code is now not longer displayed as a simple paragraph but as a
table with one single cell. A new function has been written
ODTUtility::generateODTfromHTMLCode to automatically convert HTML code
to ODT format. This can also handle list elements. The old Geshi
conversion code could not handle some elements like lists and divs.

Putting source code in a table cell allows us to insert paragraphs or
a list in the table cell. If Geshi has line numbering disabled, then we
will only get some simple text spans which will be included in a
paragraph. If Geshi has line numbering enabled, it will create an ordered
list for the line numbers. This cannot be inserted into a paragraph and
also a list cannot have a single surrounding border. Having a table
surrounding the source code gives us the border and allows for paragraph
or list content.